+++ src/DoublePinyinTable.h
@@ -32,7 +32,8 @@
 #ifdef DEFINE_DOUBLE_PINYIN_TABLES
 
 #define PINYIN_ID_AEO PINYIN_ID_ZERO
-static const gint double_pinyin_mspy_sheng[] = {
+
+static const gint8 double_pinyin_mspy_sheng[] = {
     PINYIN_ID_AEO,  // A
     PINYIN_ID_B,    // B
     PINYIN_ID_C,    // C
@@ -61,7 +62,7 @@ static const gint double_pinyin_mspy_sheng[] = {
     PINYIN_ID_Z,    // Z
     PINYIN_ID_VOID, // ;
 };
-static const gint double_pinyin_mspy_yun[][2] = {
+static const gint8 double_pinyin_mspy_yun[][2] = {
     { PINYIN_ID_A,    PINYIN_ID_VOID }, // A
     { PINYIN_ID_OU,   PINYIN_ID_VOID }, // B
     { PINYIN_ID_IAO,  PINYIN_ID_VOID }, // C
@@ -90,7 +91,7 @@ static const gint double_pinyin_mspy_yun[][2] = {
     { PINYIN_ID_EI,   PINYIN_ID_VOID }, // Z
     { PINYIN_ID_ING,  PINYIN_ID_VOID }, // ;
 };
-static const gint double_pinyin_zrm_sheng[] = {
+static const gint8 double_pinyin_zrm_sheng[] = {
     PINYIN_ID_AEO,  // A
     PINYIN_ID_B,    // B
     PINYIN_ID_C,    // C
@@ -119,7 +120,7 @@ static const gint double_pinyin_zrm_sheng[] = {
     PINYIN_ID_Z,    // Z
     PINYIN_ID_VOID, // ;
 };
-static const gint double_pinyin_zrm_yun[][2] = {
+static const gint8 double_pinyin_zrm_yun[][2] = {
     { PINYIN_ID_A,    PINYIN_ID_VOID }, // A
     { PINYIN_ID_OU,   PINYIN_ID_VOID }, // B
     { PINYIN_ID_IAO,  PINYIN_ID_VOID }, // C
@@ -148,7 +149,7 @@ static const gint double_pinyin_zrm_yun[][2] = {
     { PINYIN_ID_EI,   PINYIN_ID_VOID }, // Z
     { PINYIN_ID_VOID, PINYIN_ID_VOID }, // ;
 };
-static const gint double_pinyin_abc_sheng[] = {
+static const gint8 double_pinyin_abc_sheng[] = {
     PINYIN_ID_ZH,   // A
     PINYIN_ID_B,    // B
     PINYIN_ID_C,    // C
@@ -177,7 +178,7 @@ static const gint double_pinyin_abc_sheng[] = {
     PINYIN_ID_Z,    // Z
     PINYIN_ID_VOID, // ;
 };
-static const gint double_pinyin_abc_yun[][2] = {
+static const gint8 double_pinyin_abc_yun[][2] = {
     { PINYIN_ID_A,    PINYIN_ID_VOID }, // A
     { PINYIN_ID_OU,   PINYIN_ID_VOID }, // B
     { PINYIN_ID_IN,   PINYIN_ID_UAI  }, // C
@@ -206,7 +207,7 @@ static const gint double_pinyin_abc_yun[][2] = {
     { PINYIN_ID_IAO,  PINYIN_ID_VOID }, // Z
     { PINYIN_ID_VOID, PINYIN_ID_VOID }, // ;
 };
-static const gint double_pinyin_zgpy_sheng[] = {
+static const gint8 double_pinyin_zgpy_sheng[] = {
     PINYIN_ID_CH,   // A
     PINYIN_ID_B,    // B
     PINYIN_ID_C,    // C
@@ -235,7 +236,7 @@ static const gint double_pinyin_zgpy_sheng[] = {
     PINYIN_ID_Z,    // Z
     PINYIN_ID_VOID, // ;
 };
-static const gint double_pinyin_zgpy_yun[][2] = {
+static const gint8 double_pinyin_zgpy_yun[][2] = {
     { PINYIN_ID_A,    PINYIN_ID_VOID }, // A
     { PINYIN_ID_IAO,  PINYIN_ID_VOID }, // B
     { PINYIN_ID_VOID, PINYIN_ID_VOID }, // C
@@ -264,7 +265,7 @@ static const gint double_pinyin_zgpy_yun[][2] = {
     { PINYIN_ID_OU,   PINYIN_ID_VOID }, // Z
     { PINYIN_ID_ING,  PINYIN_ID_VOID }, // ;
 };
-static const gint double_pinyin_pyjj_sheng[] = {
+static const gint8 double_pinyin_pyjj_sheng[] = {
     PINYIN_ID_ZERO, // A
     PINYIN_ID_B,    // B
     PINYIN_ID_C,    // C
@@ -293,7 +294,7 @@ static const gint double_pinyin_pyjj_sheng[] = {
     PINYIN_ID_Z,    // Z
     PINYIN_ID_VOID, // ;
 };
-static const gint double_pinyin_pyjj_yun[][2] = {
+static const gint8 double_pinyin_pyjj_yun[][2] = {
     { PINYIN_ID_A,    PINYIN_ID_VOID }, // A
     { PINYIN_ID_IA,   PINYIN_ID_UA   }, // B
     { PINYIN_ID_UAN,  PINYIN_ID_VOID }, // C
@@ -322,7 +323,7 @@ static const gint double_pinyin_pyjj_yun[][2] = {
     { PINYIN_ID_UN,   PINYIN_ID_VOID }, // Z
     { PINYIN_ID_VOID, PINYIN_ID_VOID }, // ;
 };
-static const gint double_pinyin_xhe_sheng[] = {
+static const gint8 double_pinyin_xhe_sheng[] = {
     PINYIN_ID_AEO,  // A
     PINYIN_ID_B,    // B
     PINYIN_ID_C,    // C
@@ -351,7 +352,7 @@ static const gint double_pinyin_xhe_sheng[] = {
     PINYIN_ID_Z,    // Z
     PINYIN_ID_VOID, // ;
 };
-static const gint double_pinyin_xhe_yun[][2] = {
+static const gint8 double_pinyin_xhe_yun[][2] = {
     { PINYIN_ID_A,    PINYIN_ID_VOID }, // A
     { PINYIN_ID_IN,   PINYIN_ID_VOID }, // B
     { PINYIN_ID_AO,   PINYIN_ID_VOID }, // C
@@ -382,8 +383,8 @@ static const gint double_pinyin_xhe_yun[][2] = {
 };
 
 static const struct {
-    const gint  (&sheng)[27];
-    const gint  (&yun)[27][2];
+    const gint8  (&sheng)[27];
+    const gint8  (&yun)[27][2];
 } double_pinyin_map [] = {
     { double_pinyin_mspy_sheng, double_pinyin_mspy_yun},
     { double_pinyin_zrm_sheng, double_pinyin_zrm_yun},
